fix: resolve Wave 2 compilation errors (round 1)
Fix DataInstance<*> type params, private field access patterns,
nullable receivers, and method call syntax across 38 files.

+++ src/main/java/org/javarosa/core/model/FormIndex.kt
@@ -375,8 +375,8 @@ class FormIndex : Externalizable {
     override fun writeExternal(out: DataOutputStream) {
         ExtUtil.writeBool(out, beginningOfForm)
         ExtUtil.writeBool(out, endOfForm)
-        ExtUtil.writeNumeric(out, localIndex)
-        ExtUtil.writeNumeric(out, instanceIndex)
+        ExtUtil.writeNumeric(out, localIndex.toLong())
+        ExtUtil.writeNumeric(out, instanceIndex.toLong())
         ExtUtil.write(out, ExtWrapNullable(reference))
         ExtUtil.write(out, ExtWrapNullable(nextLevel))
     }
